Rename set_env module to env_vars
